--- a/cpp/tests/experimental/trajectory/derive_trajectories_test.cu
+++ b/cpp/tests/experimental/trajectory/derive_trajectories_test.cu
@@ -37,7 +37,8 @@
 #include <cstdint>
 
 template <typename T>
-struct DeriveTrajectoriesTest : public ::testing::Test {};
+struct DeriveTrajectoriesTest : public ::testing::Test {
+};
 
 using TestTypes = ::testing::Types<float, double>;
 TYPED_TEST_CASE(DeriveTrajectoriesTest, TestTypes);