--- a/cpp/tests/experimental/trajectory/trajectory_distances_and_speeds_test.cu
+++ b/cpp/tests/experimental/trajectory/trajectory_distances_and_speeds_test.cu
@@ -88,19 +88,19 @@ struct TrajectoryDistancesAndSpeedsTest : public ::testing::Test {
 using TestTypes = ::testing::Types<float, double>;
 TYPED_TEST_CASE(TrajectoryDistancesAndSpeedsTest, TestTypes);
 
-TYPED_TEST(TrajectoryDistancesAndSpeedsTest, OneMillionSmallTrajectories)
+TYPED_TEST(TrajectoryDistancesAndSpeedsTest, TenThousandSmallTrajectories)
 {
-  CUSPATIAL_RUN_TEST(this->run_test, 1'000'000, 50);
+  CUSPATIAL_RUN_TEST(this->run_test, 10'000, 50);
 }
 
 TYPED_TEST(TrajectoryDistancesAndSpeedsTest, OneHundredLargeTrajectories)
 {
-  CUSPATIAL_RUN_TEST(this->run_test, 100, 1'000'000);
+  CUSPATIAL_RUN_TEST(this->run_test, 100, 10'000);
 }
 
 TYPED_TEST(TrajectoryDistancesAndSpeedsTest, OneVeryLargeTrajectory)
 {
-  CUSPATIAL_RUN_TEST(this->run_test, 1, 100'000'000);
+  CUSPATIAL_RUN_TEST(this->run_test, 1, 1'000'000);
 }
 
 struct time_point_generator {